35 fail

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: rabougue <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/06/27 21:24:52 by rabougue          #+#    #+#             */
-/*   Updated: 2016/08/03 01:36:25 by rabougue         ###   ########.fr       */
+/*   Updated: 2016/08/03 06:34:51 by rabougue         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,7 +68,7 @@ void			init_struct(t_printf *printf);
 const char		*check_valid_specifier(const char *format, t_printf *print);
 const char		*countSpace(const char *format, t_printf *print);
 const char		*check_neg_sign(t_printf *print, const char *format);
-void			write_space_hex(int space, t_printf *print);
+void			write_space_hex(long space, t_printf *print);
 void			write_space_percent_s(int space, t_printf *print);
 const char		*check_space(const char *format, t_printf *print);
 const char		*precision(const char *format, int d, t_printf *print);
